# golang/reserve/service/grpc/gen/apidocs.swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "reserve/v1/api/reserve-api.proto",
    "version": "version not set"
  },
...
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/v1/warehouse/item": {
      "get": {
        "summary": "should named List but it conflicts with the List from the ReserveService in a generated openapi spec",
        "operationId": "WarehouseItemService_ItemList",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ItemListResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "tags": [
          "WarehouseItemService"
        ]
      },
      "put": {
        "operationId": "WarehouseItemService_TopUp",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1ItemTopUpResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/v1ItemTopUpRequest"
            }
          }
        ],
        "tags": [
          "WarehouseItemService"
        ]
      }
    },
    "/api/v1/warehouse/item/cost/{id}": {
      "get": {
        "operationId": "WarehouseItemService_GetItemCost",
        "responses": {
          "200": {
            "description": "A successful response.",
            "schema": {
              "$ref": "#/definitions/v1GetItemCostResponse"
            }
          },
          "default": {
            "description": "An unexpected error response.",
            "schema": {
              "$ref": "#/definitions/googlerpcStatus"
            }
          }
        },
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "tags": [
          "WarehouseItemService"
        ]
      }
    }
  },
  "definitions": {
...
    "googlerpcStatus": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "message": {
          "type": "string"
        },
        "details": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/protobufAny"
          }
        }
      }
    },
    "protobufAny": {
      "type": "object",
      "properties": {
        "@type": {
          "type": "string"
        }
      },
      "additionalProperties": {}
    },
...
    "v1GetItemCostResponse": {
      "type": "object",
      "properties": {
        "cost": {
          "type": "number",
          "format": "double"
        }
      }
    },
    "v1ItemListResponse": {
      "type": "object",
      "properties": {
        "accounts": {
          "type": "array",
          "items": {
            "type": "object",
            "$ref": "#/definitions/warehousev1Item"
          }
        }
      }
    },
    "v1ItemTopUpRequest": {
      "type": "object",
      "properties": {
        "topUp": {
          "$ref": "#/definitions/v1ItemTopUpRequestTopUp"
        }
      }
    },
    "v1ItemTopUpRequestTopUp": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "amount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
    "v1ItemTopUpResponse": {
      "type": "object",
      "properties": {
        "amount": {
          "type": "integer",
          "format": "int32"
        }
      }
    },
...
    "warehousev1Item": {
      "type": "object",
      "properties": {
        "id": {
          "type": "string"
        },
        "amount": {
          "type": "integer",
          "format": "int32"
        },
        "reserved": {
          "type": "integer",
          "format": "int32"
        },
        "updatedAt": {
          "type": "string",
          "format": "date-time"
        }
      }
    }
  }
}